+++ graphics/vk/memory/vk_mesh_assets.cpp
@@ -26,7 +26,7 @@ std::optional<std::vector<std::shared_ptr<MeshAsset>>> LoadGltfMeshes(VkRenderer
 
     std::vector<std::shared_ptr<MeshAsset>> meshes;
 
-    std::vector<uint16_t> indices;
+    std::vector<uint32_t> indices;
     std::vector<VkVertex> vertices;
 
     for (auto &[primitives, weights, name] : gltf.meshes) {
@@ -49,8 +49,8 @@ std::optional<std::vector<std::shared_ptr<MeshAsset>>> LoadGltfMeshes(VkRenderer
                 auto &indexAccessor = gltf.accessors[primitive.indicesAccessor.value()];
                 indices.reserve(indices.size() + indexAccessor.count);
 
-                fastgltf::iterateAccessor<uint16_t>(gltf, indexAccessor, [&](auto value) {
-                    indices.push_back(static_cast<uint16_t>(initialVerticesSize + value));
+                fastgltf::iterateAccessor<uint32_t>(gltf, indexAccessor, [&](auto value) {
+                    indices.push_back(static_cast<uint32_t>(initialVerticesSize + value));
                 });
             }
